+++ src/gnomon/__init__.py
@@ -9,6 +9,7 @@ from .backtesting import EvaluationBudget, evaluate_reference
 from .temporal_store import TemporalStore
 from .product_contract import __version__
 from .final_selection import forecast_completion, forecast_request_fingerprint, resolve_final_selection
+from .memory_bridge import EvidenceMemory, HermesMemoryAdapter
 
 __all__ = [
     "AdapterCapabilities", "ForecastRequest", "ForecastResult",
@@ -16,6 +17,7 @@ __all__ = [
     "TemporalLedger", "GnomonSession", "EvaluationBudget", "evaluate_reference",
     "TemporalStore", "temporal_operation", "__version__",
     "forecast_completion", "forecast_request_fingerprint", "resolve_final_selection",
+    "EvidenceMemory", "HermesMemoryAdapter",
 ]